+++ AP Lesson 14/Location.java
@@ -1,5 +1,5 @@
 public interface Location
 {
    double[] getLoc();
-   double getID();
+   int getID();
 }
